requirements.txt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
python-dotenv>=1.0.0
trafilatura>=1.6.0
lxml>=5.0.0



